% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-doc.R
\docType{package}
\name{Tcomp}
\alias{Tcomp}
\alias{Tcomp-package}
\title{Tcomp: Data from the Tourism Forecasting competition}
\description{
The Tcomp package provides data from the tourism forecasting competition described
in George Athanasopolous, Rob J. Hyndman, Haiyan Song, Doris C. Wu (2011) 
``The tourism forecasting competition'', \emph{International Journal of Forecasting} 27 (2011) 822-844.
}
\section{Tcomp data}{

\describe{
\item{tourism}{A list of class Mdata with 1,311 time series in it, divided into 
training (\code{x}) and test (\code{xx}) sets. The data are subsettable with subset method from Mcomp R package by monthly, quarterly and yearly series.}
}
}

\section{Tcomp functions}{

\describe{
\item{forecast_comp}{A convenient wrapper function for providing the mean absolute scaled error (MASE) of four common forecasting methodologies for a time series of class Mdata.}
}
}

