% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-doc.R
\docType{data}
\name{Tcomp_reproduction}
\alias{Tcomp_reproduction}
\title{Reproduction of selected tourism competition results}
\format{{A list of three elements named \code{monthly}, \code{quarterly} and \code{yearly}.   
These correspond to tables 4, 5 and 6 in the Athanasopoulos et al 2011 article.  
}}
\source{
\url{http://robjhyndman.com/papers/the-tourism-forecasting-competition}
}
\usage{
Tcomp_reproduction
}
\description{
Reproduction of selected results from the tourism forecasting competition described in 
Athanasopoulos et al. 2011 (\url{http://robjhyndman.com/papers/forecompijf.pdf})
}
\details{
Note that only Mean Absolute Percentage Error of the naive forecasts matches exactly that published.
All Mean Absolute Scaled Error results are slightly higher than those published due to an unknown
difference in MASE method.  All results for ARIMA, ETS and Theta method forecasts differ due to
changes in the forecasting methods since 2011.
See Vignette for details,
including the code required to re-create the `Tcomp_reproduction` object.
}
\examples{
Tcomp_reproduction
}
\keyword{datasets}

